@@ -3,7 +3,7 @@ package com.starsearth.two.adapter
 import android.content.Context
 import android.graphics.BitmapFactory
 import android.graphics.Color
-import android.support.v7.widget.RecyclerView
+import androidx.recyclerview.widget.RecyclerView
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
