@@ -1,8 +1,8 @@
 package com.starsearth.two.adapter
 
 import android.graphics.BitmapFactory
-import android.support.v7.widget.CardView
-import android.support.v7.widget.RecyclerView
+import androidx.cardview.widget.CardView
+import androidx.recyclerview.widget.RecyclerView
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
