@@ -2,7 +2,7 @@ package com.starsearth.two.fragments
 
 import android.content.Context
 import android.os.Bundle
-import android.support.v4.app.Fragment
+import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
@@ -31,8 +31,8 @@ class AnswerExplanationFragment : Fragment() {
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        arguments?.let {
-            response = it.getParcelable(ARG_RESPONSE)
+        arguments?.getParcelable<Response>(ARG_RESPONSE)?.let {
+            response = it
         }
     }
 
