@@ -2,10 +2,10 @@ package com.starsearth.two.fragments
 
 import android.content.Context
 import android.os.Bundle
-import android.support.v4.app.Fragment
-import android.support.v7.widget.GridLayoutManager
-import android.support.v7.widget.LinearLayoutManager
-import android.support.v7.widget.RecyclerView
+import androidx.fragment.app.Fragment
+import androidx.recyclerview.widget.GridLayoutManager
+import androidx.recyclerview.widget.LinearLayoutManager
+import androidx.recyclerview.widget.RecyclerView
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
@@ -32,8 +32,8 @@ class AutismStoryFragment : Fragment() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
 
-        arguments?.let {
-            mTask = it.getParcelable(ARG_TASK)
+        arguments?.getParcelable<Task>(ARG_TASK)?.let {
+            mTask = it
         }
     }
 
@@ -45,8 +45,15 @@ class AutismStoryFragment : Fragment() {
         if (view is RecyclerView) {
             with(view) {
                 layoutManager = when {
-                    columnCount <= 1 -> LinearLayoutManager(context, LinearLayoutManager.HORIZONTAL, false)
-                    else -> GridLayoutManager(context, columnCount)
+                    columnCount <= 1 -> LinearLayoutManager(
+                        context,
+                        LinearLayoutManager.HORIZONTAL,
+                        false
+                    )
+                    else -> GridLayoutManager(
+                        context,
+                        columnCount
+                    )
                 }
                 mAdapter = MyAutismStoryRecyclerViewAdapter(mTask.content as List<Any>, listener)
                 adapter = mAdapter
