@@ -3,7 +3,7 @@ package com.starsearth.two.fragments
 import android.content.Context
 import android.os.Bundle
 import android.os.Parcelable
-import android.support.v4.app.Fragment
+import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
@@ -35,8 +35,10 @@ class CourseDescriptionFragment : Fragment() {
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
+        arguments?.getParcelable<Course>(ARG_COURSE)?.let {
+            mCourse = it
+        }
         arguments?.let {
-            mCourse = it.getParcelable(ARG_COURSE)
             mShowTapToContinue = it.getBoolean(ARG_SHOW_TAP_TO_CONTINUE)
         }
     }
