@@ -2,7 +2,7 @@ package com.starsearth.two.fragments
 
 import android.content.Context
 import android.os.Bundle
-import android.support.v4.app.Fragment
+import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
@@ -57,9 +57,11 @@ class ResultDetailFragment : Fragment(), SeOnTouchListener.OnSeTouchListenerInte
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        arguments?.let {
-            mTask = it.getParcelable(ARG_TASK)
-            mResult = it.getParcelable(ARG_RESULT)
+        arguments?.getParcelable<Task>(ARG_TASK)?.let {
+            mTask = it
+        }
+        arguments?.getParcelable<Result>(ARG_RESULT)?.let {
+            mResult = it
         }
     }
 
