@@ -3,10 +3,10 @@ package com.starsearth.two.fragments.lists
 import android.content.Context
 import android.os.Bundle
 import android.os.Parcelable
-import android.support.v4.app.Fragment
-import android.support.v7.widget.GridLayoutManager
-import android.support.v7.widget.LinearLayoutManager
-import android.support.v7.widget.RecyclerView
+import androidx.fragment.app.Fragment
+import androidx.recyclerview.widget.GridLayoutManager
+import androidx.recyclerview.widget.LinearLayoutManager
+import androidx.recyclerview.widget.RecyclerView
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
@@ -75,10 +75,12 @@ class DetailListFragment : Fragment() {
         super.onCreate(savedInstanceState)
 
         if (arguments != null) {
-            mTeachingContent = arguments!!.getParcelable(ARG_TEACHING_CONTENT)
-            val parcelableArrayList = arguments!!.getParcelableArrayList<Parcelable>(ARG_RESULTS)
-            for (item in parcelableArrayList) {
-                mResults.add((item as Result))
+            mTeachingContent = requireArguments().getParcelable(ARG_TEACHING_CONTENT)
+            val parcelableArrayList = requireArguments().getParcelableArrayList<Parcelable>(ARG_RESULTS)
+            if (parcelableArrayList != null) {
+                for (item in parcelableArrayList) {
+                    mResults.add((item as Result))
+                }
             }
         }
     }
@@ -90,9 +92,14 @@ class DetailListFragment : Fragment() {
         if (view is RecyclerView) {
             val context = view.getContext()
             if (mColumnCount <= 1) {
-                view.layoutManager = LinearLayoutManager(context)
+                view.layoutManager =
+                    LinearLayoutManager(context)
             } else {
-                view.layoutManager = GridLayoutManager(context, mColumnCount)
+                view.layoutManager =
+                    GridLayoutManager(
+                        context,
+                        mColumnCount
+                    )
             }
             val listTitles = ArrayList<ListItem>()
             listTitles.add(ListItem.CREATOR) //Creator must be there for all tasks
@@ -181,12 +188,12 @@ class DetailListFragment : Fragment() {
     }
 
 
-    override fun onAttach(context: Context?) {
+    override fun onAttach(context: Context) {
         super.onAttach(context)
         if (context is OnTaskDetailListFragmentListener) {
             mListener = context
         } else {
-            throw RuntimeException(context!!.toString() + " must implement OnTaskDetailListFragmentListener")
+            throw RuntimeException(requireContext().toString() + " must implement OnTaskDetailListFragmentListener")
         }
     }
 
