@@ -1,6 +1,6 @@
 package com.starsearth.two.viewmodels
 
-import android.arch.lifecycle.ViewModel
+import androidx.lifecycle.ViewModel
 
 class SearchViewModel : ViewModel() {
 
